+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 * refactored `Psl\IO` handles API.
 * introduced a new `Psl\File` component.
 * refactor `Psl\Filesystem\write_file`, `Psl\Filesystem\append_file`, and `Psl\Filesystem\read_file` to use `Psl\File` component.
+* refactor `Psl\Shell\execute` to use `Psl\IO\Stream` component.